+++ src/Gate.IO.Api/Delivery/Requests/GateDeliveryPriceTriggeredOrderCancelRequest.cs
@@ -1,0 +1,9 @@
+namespace Gate.IO.Api.Delivery;
+
+/// <summary>
+/// Delivery price-triggered order cancel request
+/// </summary>
+public record GateDeliveryPriceTriggeredOrderCancelRequest
+{
+    public string Contract { get; set; }
+}