+++ src/Gate.IO.Api/Futures/Responses/GateFuturesBalance.cs
@@ -6,6 +6,12 @@ namespace Gate.IO.Api.Futures;
 public record GateFuturesBalance
 {
     /// <summary>
+    /// User ID
+    /// </summary>
+    [JsonProperty("user")]
+    public long? UserId { get; set; }
+
+    /// <summary>
     /// total is the balance after the user&#39;s accumulated deposit, withdraw, profit and loss (including realized profit and loss, fund, fee and referral rebate), excluding unrealized profit and loss.  total &#x3D; SUM(history_dnw, history_pnl, history_fee, history_refr, history_fund)
     /// </summary>
     [JsonProperty("total")]