+++ src/Gate.IO.Api/Helpers/DeliveryHelpers.cs
@@ -7,7 +7,7 @@ internal static class DeliveryHelpers
         if (string.IsNullOrEmpty(symbol))
             throw new ArgumentException("Symbol is not provided");
 
-        if (!Regex.IsMatch(symbol, "^([A-Z|0-9]{2,})_([A-Z|0-9]{2,})_([0-9]{8})$"))
-            throw new ArgumentException($"{symbol} is not a valid symbol. Should be [BaseAsset]_[QuoteAsset]_[ExpireDate], e.g. BTC_USDT_202303311");
+        if (!Regex.IsMatch(symbol, "^([A-Z|0-9]{2,})_([A-Z|0-9]{2,})(_([0-9]{8}))?$"))
+            throw new ArgumentException($"{symbol} is not a valid symbol. Should be [BaseAsset]_[QuoteAsset] or [BaseAsset]_[QuoteAsset]_[ExpireDate], e.g. BTC_USDT or BTC_USDT_20230331");
     }
 }